requirements.txt
# bot
discord.py

# env
python-dotenv

# db
asyncpg
sqlalchemy

# utility
result

# # test
# pytest
# pytest-asyncio
